refactor(gateway): extract delivery-confirm sweep wiring (#2996 P8 PR-C1)
Move the deliver-until-acked sweep wiring VERBATIM from gateway.ts into
telegram-plugin/gateway/delivery-confirm-wiring.ts: the 5 s confirm-sweep tick
body (runDeliveryConfirmSweep), the strand re-delivery
(redeliverStrandedInbound), and the #2787 Mechanism A card-suspension scoper
(sweepSuspendedTargets), behind a DeliveryConfirmWiringDeps builder.

- P0c rule preserved: the module exports the tick body; the GATEWAY still owns
  the setInterval registration (test imports stay side-effect-free — pinned by
  gateway-boot-side-effect-gating.test.ts).
- The idle gate crosses as getCurrentTurnNull ONLY (the design's C1 deps
  shape) — never the machine's eager in-turn state (the documented trap).
  Harness case 4 pins mid-turn suppression.
- Deviation from design §2 C1, flagged: NO enrolDelivery() wrapper — the two
  twin-body trackDelivery sites live inside the P7-owned handleInbound window
  (15000-17100), which this train must not touch; they stay byte-identical and
  are P7's to delete. The remaining trackDelivery callsites are unchanged.
- gateway-line-ratchet lowered 26171 -> 26100 (gateway.ts now 26066 lines).

Refs #2996. Job: reference/jobs/on-the-leash.md.
