+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-26171
+26100
